+++ src/RobotInterface/YarpImplementation/include/BipedalLocomotion/RobotInterface/YarpSensorBridge.h
@@ -224,7 +224,7 @@ public:
 
     /**
      * Get all joints' positions in radians
-     * @param[out] parameter all joints' position in radians
+     * @param[out] jointPositions all joints' position in radians
      * @param[out] receiveTimeInSeconds time at which the measurement was received
      * @warning the size is decided at the configuration and remains fixed,
      * and internal checks must be done at the implementation level by the Derived class.
@@ -247,7 +247,7 @@ public:
 
     /**
      * Get all joints' velocities in rad/s
-     * @param[out] parameter all joints' velocities in radians per second
+     * @param[out] jointVelocties all joints' velocities in radians per second
      * @param[out] receiveTimeInSeconds time at which the measurement was received
      * @warning the size is decided at the configuration and remains fixed,
      * and internal checks must be done at the implementation level by the Derived class.
@@ -256,6 +256,29 @@ public:
      */
     bool getJointVelocities(Eigen::Ref<Eigen::VectorXd> jointVelocties,
                             OptionalDoubleRef receiveTimeInSeconds = {}) final;
+
+    /**
+     * Get joint acceleration in rad/s^2
+     * @param[in] jointName name of the joint
+     * @param[out] jointAcceleration joint acceleration in radians per second squared
+     * @param[out] receiveTimeInSeconds time at which the measurement was received
+     * @return true/false in case of success/failure
+     */
+    bool getJointAcceleration(const std::string& jointName,
+                              double& jointAcceleration,
+                              OptionalDoubleRef receiveTimeInSeconds = {}) final;
+
+    /**
+     * Get all joints' accelerations in rad/s^2
+     * @param[out] jointAccelerations all joints' accelerations in radians per second squared
+     * @param[out] receiveTimeInSeconds time at which the measurement was received
+     * @warning the size is decided at the configuration and remains fixed,
+     * and internal checks must be done at the implementation level by the Derived class.
+     * This means that the user must pass a resized argument "jointAccelerations" to this method
+     * @return true/false in case of success/failure
+     */
+    bool getJointAccelerations(Eigen::Ref<Eigen::VectorXd> jointAccelerations,
+                               OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get IMU measurement
@@ -359,8 +382,8 @@ public:
      * @return true/false in case of success/failure
      */
     bool getMotorCurrent(const std::string& jointName,
-                          double& motorCurrent,
-                          OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                         double& motorCurrent,
+                         OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get all motors' currents in ampere
@@ -372,7 +395,7 @@ public:
      * @return true/false in case of success/failure
      */
     bool getMotorCurrents(Eigen::Ref<Eigen::VectorXd> motorCurrents,
-                           OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                          OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get motor PWM
@@ -382,8 +405,8 @@ public:
      * @return true/false in case of success/failure
      */
     bool getMotorPWM(const std::string& jointName,
-                          double& motorPWM,
-                          OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                     double& motorPWM,
+                     OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get all motors' PWMs
@@ -395,7 +418,7 @@ public:
      * @return true/false in case of success/failure
      */
     bool getMotorPWMs(Eigen::Ref<Eigen::VectorXd> motorPWMs,
-                           OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                      OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get motor torque in Nm
@@ -405,8 +428,8 @@ public:
      * @return true/false in case of success/failure
      */
     bool getJointTorque(const std::string& jointName,
-                          double& jointTorque,
-                          OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                        double& jointTorque,
+                        OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get all motors' torques in Nm
@@ -418,7 +441,7 @@ public:
      * @return true/false in case of success/failure
      */
     bool getJointTorques(Eigen::Ref<Eigen::VectorXd> jointTorques,
-                           OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                         OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get pid position in rad
@@ -441,7 +464,7 @@ public:
      * @return true/false in case of success/failure
      */
     virtual bool getPidPositions(Eigen::Ref<Eigen::VectorXd> pidPositions,
-                                  OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                                 OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get pid position error in rad
@@ -464,7 +487,7 @@ public:
      * @return true/false in case of success/failure
      */
     virtual bool getPidPositionErrors(Eigen::Ref<Eigen::VectorXd> pidPositionErrors,
-                                  OptionalDoubleRef receiveTimeInSeconds = {}) final;
+                                      OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
     /**
      * Get motor position in rad
@@ -479,7 +502,7 @@ public:
 
     /**
      * Get all motors' positions in rad
-     * @param[out] parameter all motors' position in radians
+     * @param[out] motorPositions all motors' position in radians
      * @param[out] receiveTimeInSeconds time at which the measurement was received
      * @warning the size is decided at the configuration and remains fixed,
      * and internal checks must be done at the implementation level by the Derived class.
@@ -502,7 +525,7 @@ public:
 
     /**
      * Get all motors' velocities in rad/s
-     * @param[out] parameter all motors' velocities in radians per second
+     * @param[out] motorVelocties all motors' velocities in radians per second
      * @param[out] receiveTimeInSeconds time at which the measurement was received
      * @warning the size is decided at the configuration and remains fixed,
      * and internal checks must be done at the implementation level by the Derived class.
@@ -511,6 +534,29 @@ public:
      */
     virtual bool getMotorVelocities(Eigen::Ref<Eigen::VectorXd> motorVelocties,
                                     OptionalDoubleRef receiveTimeInSeconds = {}) final;
+
+    /**
+     * Get motor acceleration in rad/s^2
+     * @param[in] jointName name of the joint
+     * @param[out] motorAcceleration motor acceleration in radians per second squared
+     * @param[out] receiveTimeInSeconds time at which the measurement was received
+     * @return true/false in case of success/failure
+     */
+    virtual bool getMotorAcceleration(const std::string& jointName,
+                                      double& motorAcceleration,
+                                      OptionalDoubleRef receiveTimeInSeconds = {}) final;
+
+    /**
+     * Get all motors' accelerations in rad/s
+     * @param[out] motorAccelerations all motors' accelerations in radians per second squared
+     * @param[out] receiveTimeInSeconds time at which the measurement was received
+     * @warning the size is decided at the configuration and remains fixed,
+     * and internal checks must be done at the implementation level by the Derived class.
+     * This means that the user must pass a resized argument "motorAccelerations" to this method
+     * @return true/false in case of success/failure
+     */
+    virtual bool getMotorAccelerations(Eigen::Ref<Eigen::VectorXd> motorAccelerations,
+                                       OptionalDoubleRef receiveTimeInSeconds = {}) final;
 
 private:
     /** Private implementation */