+++ src/TextLogging/src/Logger.cpp
@@ -20,23 +20,9 @@ TextLogging::Logger* const log()
 
 void TextLogging::setVerbosity(const Verbosity verbosity)
 {
-    const std::unordered_map<TextLogging::Verbosity, spdlog::level::level_enum> map{
-        {TextLogging::Verbosity::Trace, spdlog::level::level_enum::trace},
-        {TextLogging::Verbosity::Debug, spdlog::level::level_enum::debug},
-        {TextLogging::Verbosity::Info, spdlog::level::level_enum::info},
-        {TextLogging::Verbosity::Warn, spdlog::level::level_enum::warn},
-        {TextLogging::Verbosity::Err, spdlog::level::level_enum::err},
-        {TextLogging::Verbosity::Critical, spdlog::level::level_enum::critical},
-        {TextLogging::Verbosity::Off, spdlog::level::level_enum::off},
-    };
-
-    if (map.find(verbosity) == map.end())
-    {
-        log()->error("Failed to change verbosity to level {}", verbosity);
-        return;
-    }
-
-    log()->set_level(map.at(verbosity));
+    // get the verbosity underling value and convert it in spdlog enum type
+    const auto value = static_cast<std::underlying_type<Verbosity>::type>(verbosity);
+    log()->set_level(static_cast<spdlog::level::level_enum>(value));
 }
 
 } // namespace BipedalLocomotion